pyproject.toml
[tool.poetry]
name = "playwright-python-poc"
version = "0.1.0"
description = ""
authors = ["Bartek Wilczek <bwilczek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-playwright = "^0.5.2"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
src = ["tests"]

[tool.ruff.lint.pydocstyle]
convention = "google"
